Добавлен логгер, изменена структура проекта
