+++ .gitignore
@@ -1,3 +1,3 @@
 # Project exclude paths
 /.idea/
-settings/settings.py
+article_tgbot/settings/settings.py
