+++ article_tgbot/src/controller/service/admin_service.py
@@ -1,4 +1,4 @@
-from model.data_layer import DataLayer
+from article_tgbot.src.model.data_layer import DataLayer
 
 
 def is_admin(user_id):
